Solved Memory Leaked Test case
In function/method `jpeg_load_image` we check for error
while opening the file, if error is encountred then we exit the program;
Due to which we initialize the `JPEG` pointer and `comment.data` pointer
but forget to free them when we cannot load the image.
Now in this fix , `JPEG` pointer is initialized only after we sucessfully load
image file.